feat: move ErrorBoundary and make routing
