+++ src/components/main/Main.css
@@ -23,7 +23,6 @@ ul {
 .result-full .result-block {
   position: sticky;
   top: 0;
-
 }
 .result-answer {
   width: 50%;
@@ -31,7 +30,6 @@ ul {
 .result-answer .result-block {
   height: 170px;
   overflow: hidden;
-
 }
 .result-block {
   padding: 10px 10px 20px;
