+++ src/components/notfound/NotFound.tsx
@@ -1,7 +1,5 @@
 function NotFound() {
-  return (
-    <h1>Not Found 404</h1>
-  )
+  return <h1>Not Found 404</h1>;
 }
 
 export default NotFound;
